Scan all 1024 ports in sequence
Implement a 'for' loop to scan all the 1024 ports
on by one.

For the ports that open, close the established connection with
the remote server by sending a 'FIN' packet. That is done
by using the 'conn.Close()' function. It is not needed per se
but it is polite to do it.